gym-project/pyproject.toml
[tool.poetry]
name = "gym-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fuzzywuzzy = "^0.18.0"
pandas = "^1.4.2"
customtkinter = "^4.3.0"
tk = "^0.1.0"
Pillow = "^9.1.1"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
